requirements.txt
python-dotenv==0.20.0
py-cord==2.0.0